pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kconf"
version = "0.1.0"
description = "A CLI tool to easily manage your Kubernetes kubeconfig file."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT License" }
authors = [
    { name = "Antonio Maritato", email = "aptul99@gmail.com" },
]
dependencies = [
    "typer[all]",
    "pyyaml",
]

# This is the magic section that creates the command-line tool
[project.scripts]
kconf = "kconf.kube_tool:app"